+++ src/components/SocketNotifications.vue
@@ -1,13 +1,17 @@
 <template>
   <div>
-    <chat v-if="socket && isChatWidget" :socket="socket" />
+    <chat v-if="socket && chatWidget" :socket="socket" />
   </div>
 </template>
 
 <script>
 import io from "socket.io-client";
-import { mapState, mapGetters } from "vuex";
+import { mapState } from "pinia";
 import Chat from "@/components/Chat.vue";
+import { useListenerStore } from "@/store/listener-module";
+import { usePluginStore } from "@/store/plugin-module";
+import { useApplicationStore } from "@/store/application-module";
+import { useAgentStore } from "@/store/agent-module";
 
 export default {
   name: "SocketNotifications",
@@ -20,18 +24,28 @@ export default {
     };
   },
   computed: {
-    ...mapState({
-      user: (state) => state.application.user,
-      plugins: (state) => state.plugin.plugins,
-    }),
-    ...mapGetters({
-      isLoggedIn: "application/isLoggedIn",
-      isChatWidget: "application/isChatWidget",
-      socketUrl: "application/socketUrl",
-      token: "application/token",
-      isAutoSubscribeAgents: "application/isAutoSubscribeAgents",
-      subscribedAgents: "agent/subscribed",
-      agents: "agent/agents",
+    agentStore() {
+      return useAgentStore();
+    },
+    listenerStore() {
+      return useListenerStore();
+    },
+    pluginStore() {
+      return usePluginStore();
+    },
+    plugins() {
+      return this.pluginStore.plugins;
+    },
+    ...mapState(useApplicationStore, [
+      "user",
+      "isLoggedIn",
+      "chatWidget",
+      "socketUrl",
+      "token",
+    ]),
+    ...mapState(useAgentStore, {
+      subscribedAgents: "subscribed",
+      agents: "agents",
     }),
   },
   watch: {
@@ -133,7 +147,7 @@ export default {
             params: { id: data.id },
           },
         });
-        this.$store.dispatch("listener/addListener", data);
+        this.listenerStore.addListener(data);
       });
 
       this.socket.on("agents/new", (data) => {
@@ -146,7 +160,7 @@ export default {
             params: { id: data.session_id },
           },
         });
-        this.$store.dispatch("agent/addAgent", data);
+        this.agentStore.addAgent(data);
       });
 
       this.socket.on("reconnect_failed", () => {