+++ src/components/stagers/StagersTable.vue
@@ -87,12 +87,13 @@
 </template>
 
 <script>
-import { mapState } from "vuex";
 import moment from "moment";
 import DownloadMixin from "@/mixins/download-stager";
 import CopyMixin from "@/mixins/copy-stager";
 import DateTimeDisplay from "@/components/DateTimeDisplay.vue";
 import * as downloadApi from "@/api/download-api";
+import { useStagerStore } from "@/store/stager-module";
+import { useApplicationStore } from "@/store/application-module";
 
 export default {
   name: "StagersTable",
@@ -125,10 +126,18 @@ export default {
     };
   },
   computed: {
-    ...mapState({
-      stagers: (state) => state.stager.stagers,
-      stagersStatus: (state) => state.stager.status,
-    }),
+    stagerStore() {
+      return useStagerStore();
+    },
+    stagers() {
+      return this.stagerStore.stagers;
+    },
+    stagersStatus() {
+      return this.stagerStore.status;
+    },
+    user() {
+      return useApplicationStore().user.id;
+    },
     filteredStagers() {
       return this.stagers.filter((stager) => {
         if (this.onlyMyStagers) {
@@ -171,7 +180,7 @@ export default {
       this.$emit("delete-stager", item);
     },
     getStagers() {
-      this.$store.dispatch("stager/getStagers");
+      this.stagerStore.getStagers();
     },
   },
 };