+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Migrate from Vuex to Pinia
+
 ## [2.6.1] - 2023-09-25
 
 -   Make notification bell menu scrollable with a max height